GROK-17854: Diff Studio: Refactored code
Refactored the code
